updates...
git-svn-id: https://svn.code.sf.net/p/mydoggy/code/mydoggy/trunk@910 2a62bbde-051e-0410-bf3e-cfb6b34e140d

+++ mydoggy-plaf/src/main/java/org/noos/xing/mydoggy/plaf/ui/animation/TransparencyAnimation.java
@@ -33,9 +33,9 @@ public class TransparencyAnimation extends AbstractAnimation {
 
     protected float onAnimating(float animationPercent) {
         if (getAnimationDirection() == Direction.INCOMING) {
-            float animatingLengthX = (animationPercent * (1.0f - alpha));
+            double animatingLengthX = (animationPercent * (1.0f - alpha));
             synchronized(transparencyManager) {
-                transparencyManager.setAlphaModeRatio(component, 1.0f - animatingLengthX);
+                transparencyManager.setAlphaModeRatio(component, 1.0f - (float) animatingLengthX);
             }
         }
         return animationPercent;
